pantheon-main/pantheon-main/pyproject.toml
[tool.poetry]
name = "pantheon"
version = "0.0.1"
description = "Pantheon"
authors = ["Nipun Agarwal <nipun@zamp.finance>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
gunicorn = "^22.0.0"
structlog = "^24.2.0"
openai = "^1.35.14"
pytest = "^8.3.1"
anthropic = "0.47.1"
pyyaml = "^6.0.1"
pytest-cov = "^5.0.0"
diff-cover = "^9.1.1"
pandas = "^2.2.2"
dateparser = "^1.2.0"
litellm = "^1.55.3"
langfuse = "^2.57.0"
aioresponses = "^0.7.7"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.157.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
google-cloud-storage = "^2.19.0"
pypdf2 = "^3.0.1"
pdfplumber = "^0.11.5"
reportlab = "^4.2.5"
snowflake-connector-python = "^3.12.4"
snowflake-sqlalchemy = "^1.7.3"
zamp-public-workflow-sdk = "0.0.43"
watchfiles = "^1.0.4"
beautifulsoup4 = "^4.13.3"
boto3 = "^1.36.21"
pyarrow = "^19.0.0"
openpyxl = "^3.1.5"
polars = "^1.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
